refactor(background-checks): extract cancelForMember/retryForMember to bring service under 300 lines; add missing not-found test
- Extract `cancelForMember` and `retryForMember` to `background-check-retry.ts`
  following the existing `background-check-report-snapshot.ts` helper pattern,
  bringing `background-checks.service.ts` from 376 to 293 lines
- Add missing `throws when no check exists` test to the `retryForMember` suite,
  matching the parallel coverage in `cancelForMember` (NotFoundException at
  service.ts:304 was previously untested)
